requirements.txt
black==24.3.0
yapf==0.30.0
nbformat==5.10.3
isort==5.13.2
hydra-core==1.3.2
opencv-python-headless==4.8.1.78
ultralytics==8.1.29
ipykernel==6.29.4
python-dotenv==1.0.1